Struct simple_logger::SimpleLogger

pub struct SimpleLogger { /* private fields */ }

Implements Log and a set of simple builder methods for configuration.

Use the various “builder” methods on this struct to configure the logger, then call init to configure the log crate.

Implementations

impl SimpleLogger

pub fn new() -> SimpleLogger

Initializes the global logger with a SimpleLogger instance with default log level set to Level::Trace.

use simple_logger::SimpleLogger;
SimpleLogger::new().env().init().unwrap();
log::warn!("This is an example message.");

pub fn from_env() -> SimpleLogger

👎Deprecated since 1.12.0:
Use env instead. Will be removed in version 2.0.0.

Simulates env_logger behavior, which enables the user to choose log level by setting a RUST_LOG environment variable. The RUST_LOG is not set or its value is not recognized as one of the log levels, this function will use the Error level by default.

You may use the various builder-style methods on this type to configure the logger, and you must call init in order to start logging messages.

use simple_logger::SimpleLogger;
SimpleLogger::from_env().init().unwrap();
log::warn!("This is an example message.");

pub fn env(self) -> SimpleLogger

Simulates env_logger behavior, which enables the user to choose log level by setting a RUST_LOG environment variable. This will use the default level set by with_level if RUST_LOG is not set or can’t be parsed as a standard log level.

This must be called after with_level. If called before with_level, it will have no effect.

pub fn with_level(self, level: LevelFilter) -> SimpleLogger

Set the ‘default’ log level.

You can override the default level for specific modules and their sub-modules using with_module_level

This must be called before env. If called after env, it will override the value loaded from the environment.

pub fn with_module_level(self, target: &str, level: LevelFilter) -> SimpleLogger

Override the log level for some specific modules.

This sets the log level of a specific module and all its sub-modules. When both the level for a parent module as well as a child module are set, the more specific value is taken. If the log level for the same module is specified twice, the resulting log level is implementation defined.

Examples

Silence an overly verbose crate:

use simple_logger::SimpleLogger;
use log::LevelFilter;

SimpleLogger::new().with_module_level("chatty_dependency", LevelFilter::Warn).init().unwrap();

Disable logging for all dependencies:

use simple_logger::SimpleLogger;
use log::LevelFilter;

SimpleLogger::new()
    .with_level(LevelFilter::Off)
    .with_module_level("my_crate", LevelFilter::Info)
    .init()
    .unwrap();

pub fn with_target_levels(
self,
target_levels: HashMap<String, LevelFilter>
) -> SimpleLogger

👎Deprecated since 1.11.0:
Use with_module_level instead. Will be removed in version 2.0.0.

Override the log level for specific targets.

pub fn with_colors(self, colors: bool) -> SimpleLogger

Control whether messages are colored or not.

This method is only available if the colored feature is enabled.

pub fn init(self) -> Result<(), SetLoggerError>

‘Init’ the actual logger, instantiate it and configure it, this method MUST be called in order for the logger to be effective.